refactor: extract theme management to dedicated module
- Move all theme-related functions to new themeManager.js module
- Create clean separation of concerns by isolating theme functionality
- Add re-exports in utils.js for backward compatibility
- Maintain same functionality while improving code organization
